Move CLI into it's own module
