+++ src/app.js
@@ -5,7 +5,8 @@
 const { ApiPromise, WsProvider } = require('@polkadot/api');
 const PlugRuntimeTypes = require('@plugnet/plug-sdk-types');
 const testingPairs = require('@polkadot/keyring/testingPairs');
-const ArgParse = require('argparse')
+const ArgParse = require('argparse');
+const cli = require('../src/cli.js');
 
 const APP_SUCCESS = 0;
 const APP_FAIL_TRANSACTION_REJECTED = 1;
@@ -156,91 +157,8 @@ function selectSendReceiveKeypairs(keypairs){
   return [sender, receiver];
 }
 
-function forceInt(value, default_value) {
-  value = parseInt(value, 10);
-  if (isNaN(value)) {
-    value = default_value
-  }
-  return value;
-}
-
-function parseCliArguments() {
-  const default_ip_address = '127.0.0.1';
-  const default_port = 9944;
-  const default_timeout_ms = 5000;
-  const default_period_ms = 5000;
-  const default_block_delta = 10000;
-
-
-  let parser = ArgParse.ArgumentParser()
-  parser.addArgument(
-    ['--address'],
-    {
-      help: 'IP address of the node for sending transactions',
-      defaultValue: [default_ip_address, default_port],
-      metavar: ['ip-addr', 'port-num'],
-      nargs: '+',
-      dest: 'address'
-    }
-  );
-  parser.addArgument(
-    ['--timeout'],
-    {
-      help: 'Transaction timeout in ms',
-      defaultValue: default_timeout_ms,
-      metavar: 'ms',
-      nargs: '1',
-      dest: 'timeout_ms'
-    }
-  );
-  parser.addArgument(
-    ['--period'],
-    {
-      help: 'Time between transactions in ms',
-      defaultValue: default_period_ms,
-      metavar: 'ms',
-      nargs: '1',
-      dest: 'period_ms'
-    }
-  );
-  parser.addArgument(
-    ['--blocks'],
-    {
-      help: 'The change in block height required to conclude testing.',
-      defaultValue: default_block_delta,
-      metavar: 'height',
-      nargs: '1',
-      dest: 'block_delta'
-    }
-  );
-
-  let args = parser.parseArgs()
-  if (args.address.length == 1) {
-    args.address.push(default_port);
-  }
-  else if (args.address.length >= 2) {
-    args.address[1] = forceInt(args.address[1], default_port);
-  }
-
-  args.timeout_ms = forceInt(args.timeout_ms, default_timeout_ms);
-  args.period_ms = forceInt(args.period_ms, default_period_ms);
-  args.block_delta = forceInt(args.block_delta, default_block_delta);
-
-  let settings = {
-    address: `ws://${args.address[0]}:${args.address[1]}`,
-    transaction: {
-      timeout_ms: args.timeout_ms,
-      period_ms: args.period_ms
-    },
-    exit: {
-      block_delta: args.block_delta
-    }
-  }
-  return settings;
-}
-
 if (require.main === module) {
-  settings = parseCliArguments()
+  settings = cli.parseCliArguments()
   
   main(settings)
     .then(result => process.exit(result))
@@ -251,8 +169,7 @@ if (require.main === module) {
 } else {
   // Export modules for testing
   module.exports = {
-    selectSendReceiveKeypairs: selectSendReceiveKeypairs,
-    parseCliArguments: parseCliArguments
+    selectSendReceiveKeypairs: selectSendReceiveKeypairs
   }
 }
